Add "CloudWatch" to maintain backwards compatability
